+++ pkg/distributor/ingestion_rate_strategy_test.go
@@ -15,44 +15,76 @@ import (
 func TestIngestionRateStrategy(t *testing.T) {
 	t.Parallel()
 	tests := map[string]struct {
-		limits        validation.Limits
-		ring          ReadLifecycler
-		expectedLimit float64
-		expectedBurst int
+		limits                        validation.Limits
+		ring                          ReadLifecycler
+		expectedLimit                 float64
+		expectedBurst                 int
+		expectedNativeHistogramsLimit float64
+		expectedNativeHistogramsBurst int
 	}{
 		"local rate limiter should just return configured limits": {
 			limits: validation.Limits{
-				IngestionRateStrategy: validation.LocalIngestionRateStrategy,
-				IngestionRate:         float64(1000),
-				IngestionBurstSize:    10000,
+				IngestionRateStrategy:             validation.LocalIngestionRateStrategy,
+				IngestionRate:                     float64(1000),
+				IngestionBurstSize:                10000,
+				NativeHistogramIngestionRate:      float64(100),
+				NativeHistogramIngestionBurstSize: 100,
 			},
-			ring:          nil,
-			expectedLimit: float64(1000),
-			expectedBurst: 10000,
+			ring:                          nil,
+			expectedLimit:                 float64(1000),
+			expectedBurst:                 10000,
+			expectedNativeHistogramsLimit: float64(100),
+			expectedNativeHistogramsBurst: 100,
 		},
 		"global rate limiter should share the limit across the number of distributors": {
 			limits: validation.Limits{
-				IngestionRateStrategy: validation.GlobalIngestionRateStrategy,
-				IngestionRate:         float64(1000),
-				IngestionBurstSize:    10000,
+				IngestionRateStrategy:             validation.GlobalIngestionRateStrategy,
+				IngestionRate:                     float64(1000),
+				IngestionBurstSize:                10000,
+				NativeHistogramIngestionRate:      float64(100),
+				NativeHistogramIngestionBurstSize: 100,
 			},
 			ring: func() ReadLifecycler {
 				ring := newReadLifecyclerMock()
 				ring.On("HealthyInstancesCount").Return(2)
 				return ring
 			}(),
-			expectedLimit: float64(500),
-			expectedBurst: 10000,
+			expectedLimit:                 float64(500),
+			expectedBurst:                 10000,
+			expectedNativeHistogramsLimit: float64(50),
+			expectedNativeHistogramsBurst: 100,
+		},
+		"global rate limiter should handle the special case of inf ingestion rate": {
+			limits: validation.Limits{
+				IngestionRateStrategy:             validation.GlobalIngestionRateStrategy,
+				IngestionRate:                     float64(rate.Inf),
+				IngestionBurstSize:                0,
+				NativeHistogramIngestionRate:      float64(rate.Inf),
+				NativeHistogramIngestionBurstSize: 0,
+			},
+			ring: func() ReadLifecycler {
+				ring := newReadLifecyclerMock()
+				ring.On("HealthyInstancesCount").Return(2)
+				return ring
+			}(),
+			expectedLimit:                 float64(rate.Inf),
+			expectedBurst:                 0,
+			expectedNativeHistogramsLimit: float64(rate.Inf),
+			expectedNativeHistogramsBurst: 0,
 		},
 		"infinite rate limiter should return unlimited settings": {
 			limits: validation.Limits{
-				IngestionRateStrategy: "infinite",
-				IngestionRate:         float64(1000),
-				IngestionBurstSize:    10000,
+				IngestionRateStrategy:             "infinite",
+				IngestionRate:                     float64(1000),
+				IngestionBurstSize:                10000,
+				NativeHistogramIngestionRate:      float64(100),
+				NativeHistogramIngestionBurstSize: 100,
 			},
-			ring:          nil,
-			expectedLimit: float64(rate.Inf),
-			expectedBurst: 0,
+			ring:                          nil,
+			expectedLimit:                 float64(rate.Inf),
+			expectedBurst:                 0,
+			expectedNativeHistogramsLimit: float64(rate.Inf),
+			expectedNativeHistogramsBurst: 0,
 		},
 	}
 
@@ -62,6 +94,7 @@ func TestIngestionRateStrategy(t *testing.T) {
 		t.Run(testName, func(t *testing.T) {
 			t.Parallel()
 			var strategy limiter.RateLimiterStrategy
+			var nativeHistogramsStrategy limiter.RateLimiterStrategy
 
 			// Init limits overrides
 			overrides, err := validation.NewOverrides(testData.limits, nil)
@@ -71,16 +104,21 @@ func TestIngestionRateStrategy(t *testing.T) {
 			switch testData.limits.IngestionRateStrategy {
 			case validation.LocalIngestionRateStrategy:
 				strategy = newLocalIngestionRateStrategy(overrides)
+				nativeHistogramsStrategy = newLocalNativeHistogramIngestionRateStrategy(overrides)
 			case validation.GlobalIngestionRateStrategy:
 				strategy = newGlobalIngestionRateStrategy(overrides, testData.ring)
+				nativeHistogramsStrategy = newGlobalNativeHistogramIngestionRateStrategy(overrides, testData.ring)
 			case "infinite":
 				strategy = newInfiniteIngestionRateStrategy()
+				nativeHistogramsStrategy = newInfiniteIngestionRateStrategy()
 			default:
 				require.Fail(t, "Unknown strategy")
 			}
 
 			assert.Equal(t, strategy.Limit("test"), testData.expectedLimit)
 			assert.Equal(t, strategy.Burst("test"), testData.expectedBurst)
+			assert.Equal(t, nativeHistogramsStrategy.Limit("test"), testData.expectedNativeHistogramsLimit)
+			assert.Equal(t, nativeHistogramsStrategy.Burst("test"), testData.expectedNativeHistogramsBurst)
 		})
 	}
 }