+++ pkg/util/validation/validate.go
@@ -61,7 +61,8 @@ const (
 
 	// RateLimited is one of the values for the reason to discard samples.
 	// Declared here to avoid duplication in ingester and distributor.
-	RateLimited = "rate_limited"
+	RateLimited                = "rate_limited"
+	NativeHistogramRateLimited = "native_histogram_rate_limited"
 
 	// Too many HA clusters is one of the reasons for discarding samples.
 	TooManyHAClusters = "too_many_ha_clusters"